requirements.txt
Django==4.0.4
requests==2.27.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dj-rest-auth==2.2.4
pyhunter==1.7
clearbit==0.1.7